+++ postinstall.sh
@@ -15,9 +15,11 @@ ARGV2=$2 # Second argument is Plugin-Name for scipts etc.
 ARGV3=$3 # Third argument is Plugin installation folder
 ARGV5=$5 # Fifth argument is Base folder of LoxBerry
 
-/bin/sed -i "s#REPLACEBYSUBFOLDER#$ARGV3#" $ARGV5/config/plugins/$ARGV3/smartmeter.cfg
-/bin/sed -i "s#REPLACEBYNAME#$ARGV2#" $ARGV5/config/plugins/$ARGV3/smartmeter.cfg
+/bin/sed -i "s#REPLACEBYSUBFOLDER#$ARGV3#" $ARGV5/config/plugins/$ARGV3/smartmeter.json
+/bin/sed -i "s#REPLACEBYNAME#$ARGV2#" $ARGV5/config/plugins/$ARGV3/smartmeter.json
 /bin/chmod +x $ARGV5/bin/plugins/$ARGV3/vzlogger_config.pl
+/bin/chmod +x $ARGV5/bin/plugins/$ARGV3/migrate_config.pl
+/bin/chmod +x $ARGV5/bin/plugins/$ARGV3/config_value.pl
 /bin/chmod +x $ARGV5/bin/plugins/$ARGV3/vzlogger_validate.pl
 /bin/chmod +x $ARGV5/bin/plugins/$ARGV3/vzlogger_control.pl
 /bin/chmod +x $ARGV5/bin/plugins/$ARGV3/vzlogger_mqtt_bridge.pl